Update 1.9 (pre-demo version!)
-Fixed (most) ui elements as promised
-Refined dialogue sprites, with more varying body language
-Made the initial path through Arktis a bit shorter and cleaner
-Most tilesets use multi-layered rule tiles, so I can be lazier and somehow make things look even better
-Lots of new backgrounds!
-Damage no longer makes you lose time immediately; instead, your hourglass begins to break, and once it does you will start to lose time faster
-Major story changes, that unfortunately have not been applied to all characters
-Fisher is Scottish now

***

Future (needed) fixes:
- Crab boss is trivial, add shockwaves
- Cleaner transition into the Shore Storm
- Make it more obvious when the hourglass shatters; animation perhaps
- Sound effects. As soon as possible.

+++ Bear Witness/Assets/Scripts/Data Types/PlayerData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [System.Serializable]
 public class PlayerData
 {
-    public int playerMaxHealth = 5;
+    public float playerMaxHealth = 5;
     public List<Item> currentItems;
     public List<Item> inventory;
     public int inventoryNextSpace = 0;
